requirements.txt
PyQt5>=5.15.0
opencv-python>=4.5.0
Pillow>=8.0.0
pytesseract>=0.3.0
openai>=1.0.0
keyboard>=0.13.5
pywin32>=300.0; platform_system=="Windows"
numpy>=1.19.0
psutil>=5.8.0
python-xlib>=0.31; platform_system=="Linux"
pyobjc>=7.0; platform_system=="Darwin" 